Add velocity tests
The L_inf error on the 2D and 3D cases is a good quick check for bugs, but it shows the convergence in 3D towards the analytic solution is poor. This is probably because the vorticity in that cases is discontinuous. We might need to determine an alternative case for the paper.