Improved residual fix
There is a bit of a bug in `residual!` which was causing issues when MG levels was very big. I switched to `fix_resid!` which is also better since it skips a `mult` we don't need (since `p=0`) and fixes the residual only on the boundaries, which is where the error comes from.