Finish 2D testing
I've set the 2D dis=4 instead of 7 (like we did in the paper) since the fmm interpolation error overwhelms the reduction. The 2D and 3D error levels match now: in both cases, the BC error is much smaller with the tree sum, but it doesn't seem to be reflected in any change in the initial simulation accuracy.